refactor: mount all wire protocol primitives on Connection
We no longer use the `makeServerTrampoline` trick to pass a connection
into the wire protocol primitives. Instead, the primitives are now
mounted directly onto a `Connection`. An additional refactor is
included which specializes `Connection` for auto encryption in the
form of a `CryptoConnection`.

NODE-2133
