+++ src/cmap/auth/scram.ts
@@ -2,7 +2,7 @@ import * as crypto from 'crypto';
 import { Binary, Document } from '../../bson';
 import { MongoError, AnyError } from '../../error';
 import { AuthProvider, AuthContext } from './auth_provider';
-import type { Callback } from '../../utils';
+import { Callback, ns } from '../../utils';
 import type { MongoCredentials } from './mongo_credentials';
 import type { HandshakeDocument } from '../connect';
 
@@ -112,7 +112,7 @@ function executeScram(cryptoMethod: CryptoMethod, authContext: AuthContext, call
   const db = credentials.source;
 
   const saslStartCmd = makeFirstMessage(cryptoMethod, credentials, nonce);
-  connection.command(`${db}.$cmd`, saslStartCmd, (_err, result) => {
+  connection.command(ns(`${db}.$cmd`), saslStartCmd, undefined, (_err, result) => {
     const err = resolveError(_err, result);
     if (err) {
       return callback(err);
@@ -198,7 +198,7 @@ function continueScramConversation(
     payload: new Binary(Buffer.from(clientFinal))
   };
 
-  connection.command(`${db}.$cmd`, saslContinueCmd, (_err, r) => {
+  connection.command(ns(`${db}.$cmd`), saslContinueCmd, undefined, (_err, r) => {
     const err = resolveError(_err, r);
     if (err) {
       return callback(err);
@@ -220,7 +220,7 @@ function continueScramConversation(
       payload: Buffer.alloc(0)
     };
 
-    connection.command(`${db}.$cmd`, retrySaslContinueCmd, callback);
+    connection.command(ns(`${db}.$cmd`), retrySaslContinueCmd, undefined, callback);
   });
 }
