--- src/cmap/wire_protocol/index.ts
@@ -1,4 +1,0 @@
-export { killCursors } from './kill_cursors';
-export { getMore } from './get_more';
-export { query } from './query';
-export { command } from './command';
