+++ test/functional/aggregation.test.js
@@ -54,8 +54,6 @@ describe('Aggregation', function () {
         var collection = db.collection('shouldCorrectlyExecuteSimpleAggregationPipelineUsingArray');
         // Insert the docs
         collection.insertMany(docs, { w: 1 }, function (err, result) {
-          if (err) console.dir(err);
-
           expect(result).to.exist;
           expect(err).to.not.exist;
