+++ test/functional/cmap/connection.test.js
@@ -20,7 +20,7 @@ describe('Connection', function () {
       expect(err).to.not.exist;
       this.defer(_done => conn.destroy(_done));
 
-      conn.command('admin.$cmd', { ismaster: 1 }, (err, ismaster) => {
+      conn.command('admin.$cmd', { ismaster: 1 }, undefined, (err, ismaster) => {
         expect(err).to.not.exist;
         expect(ismaster).to.exist;
         expect(ismaster.ok).to.equal(1);
@@ -44,7 +44,7 @@ describe('Connection', function () {
       conn.on('commandSucceeded', event => events.push(event));
       conn.on('commandFailed', event => events.push(event));
 
-      conn.command('admin.$cmd', { ismaster: 1 }, (err, ismaster) => {
+      conn.command('admin.$cmd', { ismaster: 1 }, undefined, (err, ismaster) => {
         expect(err).to.not.exist;
         expect(ismaster).to.exist;
         expect(ismaster.ok).to.equal(1);
@@ -85,33 +85,35 @@ describe('Connection', function () {
           test: Math.floor(Math.random() * idx)
         }));
 
-        conn.command(ns, { insert: 'test', documents }, (err, res) => {
-          expect(err).to.not.exist;
-          expect(res).nested.property('n').to.equal(documents.length);
-
-          let totalDocumentsRead = 0;
-          conn.command(ns, { find: 'test', batchSize: 100 }, (err, result) => {
+        conn.command(ns, { drop: 'test' }, undefined, () => {
+          conn.command(ns, { insert: 'test', documents }, undefined, (err, res) => {
             expect(err).to.not.exist;
-            expect(result).nested.property('cursor').to.exist;
-            const cursor = result.cursor;
-            totalDocumentsRead += cursor.firstBatch.length;
+            expect(res).nested.property('n').to.equal(documents.length);
 
-            conn.command(
-              ns,
-              { getMore: cursor.id, collection: 'test', batchSize: 100 },
-              { exhaustAllowed: true },
-              (err, result) => {
-                expect(err).to.not.exist;
-                expect(result).nested.property('cursor').to.exist;
-                const cursor = result.cursor;
-                totalDocumentsRead += cursor.nextBatch.length;
+            let totalDocumentsRead = 0;
+            conn.command(ns, { find: 'test', batchSize: 100 }, undefined, (err, result) => {
+              expect(err).to.not.exist;
+              expect(result).nested.property('cursor').to.exist;
+              const cursor = result.cursor;
+              totalDocumentsRead += cursor.firstBatch.length;
 
-                if (cursor.id === 0 || cursor.id.isZero()) {
-                  expect(totalDocumentsRead).to.equal(documents.length);
-                  done();
+              conn.command(
+                ns,
+                { getMore: cursor.id, collection: 'test', batchSize: 100 },
+                { exhaustAllowed: true },
+                (err, result) => {
+                  expect(err).to.not.exist;
+                  expect(result).nested.property('cursor').to.exist;
+                  const cursor = result.cursor;
+                  totalDocumentsRead += cursor.nextBatch.length;
+
+                  if (cursor.id === 0 || cursor.id.isZero()) {
+                    expect(totalDocumentsRead).to.equal(documents.length);
+                    done();
+                  }
                 }
-              }
-            );
+              );
+            });
           });
         });
       });
